basic makefile and main.c
