+++ Makefile
@@ -1,0 +1,4 @@
+all:
+	cc -Wall -Wextra -Werror -lreadline *.c -o minishell
+clean:
+	rm minishell
