+++ minishell.h
@@ -6,5 +6,10 @@
 #include <stdlib.h>
 #include <dirent.h>
 #include <sys/wait.h>
+#include <readline/readline.h>
+#include <readline/history.h>
+
+int	ft_strcmp(const char *s1, const char *s2);
+int my_echo(int ac, char **av);
 
 # endif
